graphene-backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
aniso8601==7.0.0
    # via graphene
backoff==2.2.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
blinker==1.6.2
    # via flask
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via flask
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
flask==2.3.2
    # via
    #   -r requirements.in
    #   flask-graphql
flask-graphql==2.0.1
    # via -r requirements.in
googleapis-common-protos==1.59.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
graphene==2.1.9
    # via
    #   -r requirements.in
    #   graphene-federation
graphene-federation==0.2.0
    # via -r requirements.in
graphql-core==2.3.2
    # via
    #   flask-graphql
    #   graphene
    #   graphql-relay
    #   graphql-server-core
graphql-relay==2.0.1
    # via graphene
graphql-server-core==1.2.0
    # via flask-graphql
grpcio==1.54.2
    # via opentelemetry-exporter-otlp-proto-grpc
idna==3.4
    # via requests
importlib-metadata==6.0.1
    # via opentelemetry-api
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
opentelemetry-api==1.18.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-exporter-otlp==1.18.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-common==1.18.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.18.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.18.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.39b0
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-flask==0.39b0
    # via -r requirements.in
opentelemetry-instrumentation-wsgi==0.39b0
    # via opentelemetry-instrumentation-flask
opentelemetry-proto==1.18.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.18.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.39b0
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.39b0
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-wsgi
promise==2.3
    # via
    #   graphql-core
    #   graphql-relay
    #   graphql-server-core
protobuf==4.23.2
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
requests==2.31.0
    # via opentelemetry-exporter-otlp-proto-http
rx==1.6.3
    # via graphql-core
six==1.16.0
    # via
    #   graphene
    #   graphql-core
    #   graphql-relay
    #   promise
typing-extensions==4.6.3
    # via opentelemetry-sdk
urllib3==2.0.2
    # via requests
werkzeug==2.3.4
    # via flask
wrapt==1.15.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
